Got the food item form mostly visually working and mostly pretty.

+++ DatabaseAccess/Entities/FoodItemNutrient.cs
@@ -3,6 +3,6 @@ namespace SnackAndTrack.DatabaseAccess.Entities {
         public required Guid Id { get; set; }
         public required FoodItem FoodItem { get; set; }
         public required Nutrient Nutrient { get; set; }
-        public required Int16 Quantity { get; set; }
+        public required Single Quantity { get; set; }
     }
 }
